rearrange some driver tests
